Tools API tweaks, merge into lmTools
