+++ src/vscode-dts/vscode.proposed.lmTools.d.ts
@@ -13,7 +13,7 @@ declare module 'vscode' {
 	export interface LanguageModelChatFunction {
 		name: string;
 		description: string;
-		parametersSchema: JSONSchema;
+		parametersSchema?: JSONSchema;
 	}
 
 	// API -> LM: add tools as request option
@@ -54,5 +54,35 @@ declare module 'vscode' {
 
 	export interface LanguageModelChatMessage {
 		content2: string | LanguageModelChatMessageFunctionResultPart;
+	}
+
+	// Tool registration/invoking between extensions
+
+	export namespace lm {
+		/**
+		 * Register a LanguageModelTool. The tool must also be registered in the package.json `languageModelTools` contribution point.
+		 */
+		export function registerTool(name: string, tool: LanguageModelTool): Disposable;
+
+		/**
+		 * A list of all available tools.
+		 */
+		export const tools: ReadonlyArray<LanguageModelToolDescription>;
+
+		/**
+		 * Invoke a tool with the given parameters.
+		 */
+		export function invokeTool(name: string, parameters: Object, token: CancellationToken): Thenable<string>;
+	}
+
+	// Is the same as LanguageModelChatFunction now, but could have more details in the future
+	export interface LanguageModelToolDescription {
+		name: string;
+		description: string;
+		parametersSchema?: JSONSchema;
+	}
+
+	export interface LanguageModelTool {
+		invoke(parameters: any, token: CancellationToken): Thenable<string>;
 	}
 }
